requirements.txt
colorlog==6.6.0
tabulate==0.8.10
requests==2.28.0
google-cloud-pubsub==2.13.2
google-cloud-compute==1.3.2
google-cloud-monitoring==2.10.0
google_cloud_bigquery==2.34.3
google-api-python-client==2.52.0

docker==5.0.3
